pyproject.toml
[tool.poetry]
name = "weather-website"
version = "0.1.0"
description = ""
authors = ["Andrey-Barinov <barbpro89@gmail.com>"]
readme = "README.md"
packages = [
    { include = "weather_website" },
]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
geopy = "^2.4.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
flake8 = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
